awssam/ideablog/requirements.txt
aadict==0.2.3
aioredis==1.0.0
alembic==1.1.0.dev0
amqp==2.6.1
aniso8601==3.0.2
ansible==2.6.3
anyjson==0.3.3
appdirs==1.4.3
apturl==0.5.2
argh==0.26.2
argon2-cffi==20.1.0
asgi-redis==1.4.3
asgiref==3.2.10
asn1crypto==0.24.0
asset==0.6.12
astroid==2.4.2
async-generator==1.10
async-timeout==2.0.0
attrs==20.2.0
autobahn==18.4.1
Automat==0.6.0
autopep8==1.5.4
Babel==2.5.3
backcall==0.1.0
bcrypt==3.1.4
beautifulsoup4==4.9.3
billiard==3.6.3.0
bleach==3.2.1
blinker==1.4
Brlapi==0.7.0
cachetools==4.0.0
celery==4.4.7
certifi==2018.11.29
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
daphne==2.5.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Delorean==1.0.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.43.0
django-bootstrap3==14.2.0
django-bootstrap4==2.3.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-contrib-comments==1.9.2
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-filter==2.4.0
django-grappelli==2.14.2
django-guardian==2.3.0
django-heroku==0.3.1
django-jenkins==0.110.0
django-jet==1.0.8
django-js-asset==1.2.2
django-mptt==0.11.0
django-pure-pagination==0.3.0
django-simpleui==2020.9.26
django-timezone-field==4.0
django-tinymce==3.1.0
djangorestframework==3.12.1
dnf==0.0.1
dnspython==2.0.0
docutils==0.16
dramatiq==1.9.0
dukpy==0.2.3
duplicity==0.8.12.0
email-validator==1.1.1
entrypoints==0.3
fasteners==0.14.1
feedparser==6.0.1
filebrowser-safe==0.5.0
filelock==3.0.12
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
future==0.18.2
gcloud==0.17.0
gevent==20.9.0
globre==0.1.5
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
gpg===1.13.1-unknown
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
grappelli-safe==0.5.2
greenlet==0.4.17
gunicorn==20.0.4
hiredis==1.1.0
httplib2==0.14.0
humanize==3.0.1
hyperlink==20.0.1
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
Interactive==1.0.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.5.4
itsdangerous==1.1.0
itypes==1.2.0
javascripthon==0.11
jedi==0.17.2
Jinja2==2.10.1
jsonfield==3.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jws==0.1.3
keyring==18.0.1
kombu==4.6.11
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
ldap3==2.4.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
macropy3==1.1.0b2
Mako==1.1.0
Markdown==3.3
MarkupSafe==1.1.0
mccabe==0.6.1
Mezzanine==4.3.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.0
msgpack-python==0.5.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
netifaces==0.10.4
notebook==6.1.4
oauth2client==3.0.0
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
packaging==20.4
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.7.1
passlib==1.7.2
pathtools==0.1.2
pbr==5.5.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==11.9.0
prettytable==1.0.1
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.6.1
psutil==5.5.1
psycopg2==2.8.5
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.4.3
pycups==1.9.73
PyDrive==1.3.1
pyecharts==1.8.1
pyecharts-javascripthon==0.0.6
pyecharts-jupyter-installer==0.0.3
Pygments==2.7.1
PyGObject==3.36.0
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyMySQL==0.10.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyQt5==5.14.1
Pyrebase==3.0.27
pyRFC3339==1.1
pyrsistent==0.17.3
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.1
python-crontab==2.5.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-instagram==1.3.2
python-jwt==2.0.1
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
readme-renderer==28.0
redis==2.10.6
reportlab==3.5.34
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
requests-unixsocket==0.2.0
rsa==4.0
Rx==1.6.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
sgmllib3k==1.0.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
sip==4.19.21
six==1.14.0
soupsieve==2.0.1
speaklater==1.4
SQLAlchemy==1.3.12
sqlparse==0.4.0
sshtunnel==0.1.4
systemd-python==234
terminado==0.9.1
testpath==0.4.4
testresources==2.0.1
toml==0.10.1
torbrowser-launcher==0.3.2
tornado==6.0.4
traitlets==5.0.5
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.1
uritemplate==0.6
urllib3==1.25.8
usb-creator==0.3.7
uWSGI==2.0.19.1
vine==1.3.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
watchdog==0.10.3
watchdog-gevent==0.1.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.2.1
xkit==0.0.0
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.1.2
